feat(runtime): extract /automation dispatcher domain body — ADR-0076 D11 step ③ PR-6 (#2462)
The automation service bridge (~260 lines) moves to
domains/automation.ts. Cleanest extraction of the series: zero new
DomainHandlerDeps growth (getService/resolveService/success/error cover
it). Route-order subtlety preserved verbatim — /actions, /connectors
and /_status keep their guard positions before the /:name→getFlow
catch-all (a flow literally named "actions" must not shadow them; a new
seam test locks the guard position).

NOTE: /actions (handleActions) is NOT in this cut — it is
deep-coupled to env-resolution state (this.kernel / kernelResolver /
resolveDefaultProject + 4 private helpers) and goes with the /data
/meta deep-coupling batch.

Verified: seam suite 37 tests, runtime 642 green, http-conformance 41
green, dependent closure builds with DTS (--force).
